neuro/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-login==0.6.2
flask-wtf==1.0.1
python-dotenv==1.0.0
flask-bootstrap==3.3.7.1
email-validator==1.3.1